Students' dataset added
I do not own the dataset. It is located in the following url: https://archive.ics.uci.edu/ml/datasets/Student+Performance However, in order to facilitate the download and setup of the project, it has been added to the repository.
The command setwd has been removed in order to make the script work in any computer